Move config class to its own file
